Изменения по таблицам раздела обо мне.

+++ Profile/pgsql.leoka_estetica_dev/Tables/[Profile].[UserIntents]/[Profile].[UserIntents]_CreateTable.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS "Profile"."UserIntents"
 (
-    "UserIntentId" SERIAL,
+    "UserIntentId" BIGSERIAL,
     "IntentId" INT NOT NULL,
     "UserId" BIGINT NOT NULL,
     "Position" INT NOT NULL,
